B-因子构建类/个股动量效应的识别及球队硬币因子/requirements.txt
alphalens==0.4.0
empyrical==0.5.5
fire==0.5.0
joblib==1.2.0
lightgbm==3.3.5
loguru==0.6.0
matplotlib==3.7.1
numpy==1.23.5
pandas==1.5.3
pyqlib==0.9.3
python_dateutil==2.8.2
rich==13.4.2
scipy==1.10.1
seaborn==0.12.2
SQLAlchemy==1.4.49
statsmodels==0.13.5
tqdm==4.65.0
